update object store list
